added mocha for stubbing/mocking, added venue controller test basic
