+++ app/controllers/users_controller.rb
@@ -22,7 +22,7 @@ class UsersController < ApplicationController
     @user.password = pass if pass == params[:password_conf]
 
     if @user.save
-      redirect_to :index
+      render :index
     else
       # TODO: could probably change the flash display to enumerate @user.errors
       flash.now[:error] = "Something went wrong - please check your fields and try again!"
